pyproject.toml
[project]
name = "armarinho-tpv-bot"
version = "0.1.0"
description = "Bot de geração de QR code para uso nos armários do TPV"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pillow>=11.1.0",
  "python-dotenv>=1.0.1",
  "python-telegram-bot>=21.10",
  "qrcode>=8.0",
  "ruff>=0.9.7",
]
